requirements.txt
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
google-auth==2.48.0
google-auth-oauthlib==1.2.4
gspread==6.2.1
httplib2==0.31.2
icalendar==7.0.2
idna==3.11
oauth2client==4.1.3
oauthlib==3.3.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pyparsing==3.3.2
pyTelegramBotAPI==4.31.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
schedule==1.2.2
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
